Find duplicates completed.

+++ CPP_Codes/findFactorial.cpp
@@ -1,8 +1,10 @@
 #include<iostream>
+#include <iomanip>
+#include<boost/multiprecision/cpp_int.hpp>
 using namespace std;
 
 int main(int argc, char const *argv[]) {
-  long double f = 1;
+  cpp_int f = 1;
   int n, i;
   cin>>n;
   for(i=1; i<=n; i++) {
